Export the URL generator
